Versión 1.5

+++ capturaPosturas.py
@@ -9,7 +9,26 @@ def findAngle(x1, y1, x2, y2):
     degree = int(180 / m.pi) * theta
     return degree
 
-def captura():
+'''
+accion 1.- Update
+category 1.- Good 2.- Regular 3.- Bad
+filename: filename path 
+'''
+
+def count_files_by_extension(dirname, extension):
+    # Obtener la lista de archivos en la carpeta
+    files = os.listdir(dirname)
+
+    # Filtrar archivos por la extensión deseada
+    files_with_extension = [file for file in files if file.endswith(extension)]
+
+    # Contar la cantidad de archivos con la extensión
+    amount_files = len(files_with_extension)
+
+    return amount_files
+
+
+def CapturePosture(category:int, filename:list):
 
     # Font type.
     font = cv2.FONT_HERSHEY_SIMPLEX
@@ -23,10 +42,8 @@ def captura():
     pose = mp_pose.Pose()
 
     # Colors.
-    blue = (255, 127, 0)
     red = (50, 50, 255)
     green = (127, 255, 0)
-    dark_blue = (127, 20, 0)
     light_green = (127, 233, 100)
     yellow = (0, 255, 255)
     pink = (255, 0, 255)
@@ -37,18 +54,18 @@ def captura():
     if not os.path.exists(data):
         os.makedirs(data)
 
-    personpath = os.path.join(data, 'Entrenamiento')
-    valpath = os.path.join(data, 'Validacion')
-    goodposture = os.path.join(personpath, 'Buena')
-    val_good = os.path.join(valpath, 'Buena')
-    badposture = os.path.join(personpath, 'Mala')
-    val_bad = os.path.join(valpath, 'Mala')
-    regularposture = os.path.join(personpath, 'Regular')
-    val_regular = os.path.join(valpath, 'Regular')    
+    personpath = os.path.join(data, 'Training')
+    valpath = os.path.join(data, 'Validation')
+    goodposture = os.path.join(personpath, 'Good_Posture')
+    val_good = os.path.join(valpath, 'Good_Posture')
+    badposture = os.path.join(personpath, 'Bad_Posture')
+    val_bad = os.path.join(valpath, 'Bad_Posture')
+    regularposture = os.path.join(personpath, 'Regular_Posture')
+    val_regular = os.path.join(valpath, 'Regular_Posture')    
 
     if  os.path.exists(personpath) and os.path.exists(valpath):
-        print(f"Carpeta encontrada: {personpath} y {valpath}")
-        
+        #print(f"File successfully finding: {personpath} y {valpath}")
+        pass
     else:
         os.makedirs(personpath)
         os.makedirs(valpath)
@@ -61,17 +78,19 @@ def captura():
         
         os.makedirs(regularposture)
         os.makedirs(val_regular)
-
+    '''///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////'''
+    
     cap = cv2.VideoCapture(0, cv2.CAP_DSHOW)
     
-    count = 0
-    good_frames = 0
-    bad_frames = 0
-    regular_frames = 0
-    total_good_postures = 0
-    total_bad_postures = 0
-    total_regular_posture = 0
     
+    good_frames = 1
+    bad_frames = 1
+    regular_frames = 1
+    total_good_postures = 1
+    total_bad_postures = 1
+    total_regular_posture = 1
+    
+    # limite
     while True:
         success, image = cap.read()
         if not success:
@@ -108,8 +127,9 @@ def captura():
 
             angle_text_string = 'Neck : ' + str(int(neck_inclination)) + '  Torso : ' + str(int(torso_inclination))
 
-            if  neck_inclination < 20 and neck_inclination > 0 and torso_inclination < 5 and torso_inclination > 0 :
+            if  neck_inclination < 20 and neck_inclination > 0 and torso_inclination < 5 and torso_inclination > 0  and count_files_by_extension(goodposture, 'jpg') < 250:
                 good_frames += 1
+
                 cv2.putText(image, angle_text_string, (10, 30), font, 0.9, light_green, 2)
                 cv2.putText(image, str(int(neck_inclination)), (l_shldr_x + 10, l_shldr_y), font, 0.9, light_green, 2)
                 cv2.putText(image, str(int(torso_inclination)), (l_hip_x + 10, l_hip_y), font, 0.9, light_green, 2)
@@ -117,18 +137,27 @@ def captura():
                 cv2.line(image, (l_shldr_x, l_shldr_y), (l_shldr_x, l_shldr_y - 100), green, 4)
                 cv2.line(image, (l_hip_x, l_hip_y), (l_shldr_x, l_shldr_y), green, 4)
                 cv2.line(image, (l_hip_x, l_hip_y), (l_hip_x, l_hip_y - 100), green, 4)
-                img_filename = os.path.join(goodposture, 'good_{}.jpg'.format(good_frames))
-                
-                if not os.path.exists(img_filename):
+
+                if category == 1: # Good update
+                    img_filename = os.path.join(goodposture, filename)
+                else: # Insert
+                    img_filename = os.path.join(goodposture, 'good_{}.jpg'.format(total_good_postures))
+
+                if not os.path.exists(img_filename) and category == 1: # Update
                     cv2.imwrite(img_filename, image)
-                                   
+                    break
+                elif not os.path.exists(img_filename): # Insert
+                    cv2.imwrite(img_filename, image)
+                
                 if good_frames<11:#se toman de las primeras 10 imagenes que se capturan de la camara
-                    img_val_filename = os.path.join(val_good, 'good_{}.jpg'.format(good_frames))
-                    cv2.imwrite(img_val_filename, image)
+                    img_val_filename = os.path.join(val_good, 'good_{}.jpg'.format(total_good_postures))
+                    if not os.path.exists(img_val_filename):
+                        cv2.imwrite(img_val_filename, image)
+                
                 # Cada vez que se detecta una buena postura, aumenta el contador de buenas posturas
                 total_good_postures += 1
                 
-            elif neck_inclination < 40 and neck_inclination > 20 and torso_inclination < 10 and torso_inclination > 5:
+            elif neck_inclination < 40 and neck_inclination > 20 and torso_inclination < 10 and torso_inclination > 5 and count_files_by_extension(regularposture, 'jpg') < 250:
 
                 regular_frames += 1
                 cv2.putText(image, angle_text_string, (10, 30), font, 0.9, yellow, 2)
@@ -138,17 +167,28 @@ def captura():
                 cv2.line(image, (l_shldr_x, l_shldr_y), (l_shldr_x, l_shldr_y - 100), yellow, 4)
                 cv2.line(image, (l_hip_x, l_hip_y), (l_shldr_x, l_shldr_y), yellow, 4)
                 cv2.line(image, (l_hip_x, l_hip_y), (l_hip_x, l_hip_y - 100), yellow, 4)
-                img_filename = os.path.join(regularposture, 'regular_{}.jpg'.format(regular_frames))
                 
-                if not os.path.exists(img_filename):
+                if category == 2: # Regular
+                    img_filename = os.path.join(regularposture, filename)
+                else: # Insertar
+                    img_filename = os.path.join(regularposture, 'regular_{}.jpg'.format(total_regular_posture))
+                
+
+                if not os.path.exists(img_filename) and category == 2: # Actualizar
                     cv2.imwrite(img_filename, image)
+                    break
+                elif not os.path.exists(img_filename): # Insert
+                    cv2.imwrite(img_filename, image)
+
                     
                 if regular_frames<11:#se toman de las primeras 10 imagenes que se capturan de la camara
-                    img_val_filename = os.path.join(val_regular, 'regular_{}.jpg'.format(regular_frames))
-                    cv2.imwrite(img_val_filename, image)
+                    img_val_filename = os.path.join(val_regular, 'regular_{}.jpg'.format(total_regular_posture))
+                    if not os.path.exists(img_val_filename):
+                        cv2.imwrite(img_val_filename, image)
+                
                 total_regular_posture += 1
                 
-            else:
+            elif count_files_by_extension(badposture, 'jpg') < 250:
                   
                 bad_frames += 1
                 cv2.putText(image, angle_text_string, (10, 30), font, 0.9, red, 2)
@@ -158,26 +198,43 @@ def captura():
                 cv2.line(image, (l_shldr_x, l_shldr_y), (l_shldr_x, l_shldr_y - 100), red, 4)
                 cv2.line(image, (l_hip_x, l_hip_y), (l_shldr_x, l_shldr_y), red, 4)
                 cv2.line(image, (l_hip_x, l_hip_y), (l_hip_x, l_hip_y - 100), red, 4)
-                img_filename = os.path.join(badposture, 'bad_{}.jpg'.format(bad_frames))
                 
-                if not os.path.exists(img_filename):
+                if category == 3: # Bad
+                    img_filename = os.path.join(badposture, filename)
+                else: # Insertar
+                    img_filename = os.path.join(badposture, 'bad_{}.jpg'.format(total_bad_postures))
+
+
+                if not os.path.exists(img_filename) and category == 3: # Actualizar
+                    cv2.imwrite(img_filename, image)
+                    break
+                elif not os.path.exists(img_filename): # Insertar
                     cv2.imwrite(img_filename, image)
                 
+                
                 if bad_frames<11:#se toman de las primeras 10 imagenes que se capturan de la camara
-                    img_val_filename = os.path.join(val_bad, 'bad_{}.jpg'.format(bad_frames))
-                    cv2.imwrite(img_val_filename, image)
+                    img_val_filename = os.path.join(val_bad, 'bad_{}.jpg'.format(total_bad_postures))
+                    if not os.path.exists(img_val_filename):
+                        cv2.imwrite(img_val_filename, image)
                 # Cada vez que se detecta una mala postura, aumenta el contador de malas posturas
                 total_bad_postures += 1
-
-        count += 1
+            print(count_files_by_extension(goodposture, 'jpg'))
+            print(count_files_by_extension(regularposture, 'jpg'))
+            print(count_files_by_extension(badposture, 'jpg'))
         # Display the image
         cv2.imshow('frame', image)
-
-        if cv2.waitKey(1) & 0xFF == ord('q') :  # Press 'q' to exit the loop
+        print(count_files_by_extension(goodposture, 'jpg'))
+        print(count_files_by_extension(regularposture, 'jpg'))
+        print(count_files_by_extension(badposture, 'jpg'))
+        #Paro generico para insercion de frames
+        if count_files_by_extension(goodposture, 'jpg') == 250 and count_files_by_extension(regularposture, 'jpg') == 250 and count_files_by_extension(badposture, 'jpg') == 250 or cv2.waitKey(1) & 0xFF == ord('q'):
+            cap.release()
+            cv2.destroyAllWindows()
             break
 
-    cap.release()
-    cv2.destroyAllWindows()
+    
 
 if __name__ == '__main__':
-    captura()
+    category = 0
+    filename = ''
+    CapturePosture(category, filename)
